refactor: move auth components to folder
